Np 45462 nvi approval reset on affiliation change (#4941)
* added affiliation id count dictionary

* filter on affiliation.id

* moved nvi approval reset checks to nviHelpers

* updatd function to loop through contributors and checkfor affiliation changes

* included check for contributirs in affiliations check, and added cache for topLevelOrgs

* removed function call + code cleanup

* removed fallback
